--- a/test/triangular3.jl
+++ b/test/triangular3.jl
@@ -1,6 +1,6 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-module TestTriangularReal
+module TestTriangularComplex
 
 using Test, LinearAlgebra, Random
 using LinearAlgebra: BlasFloat, errorbounds, full!, transpose!
@@ -9,6 +9,6 @@ Random.seed!(123)
 
 include("testtriag.jl") # test_approx_eq_modphase
 
-test_triangular((Float32, Float64, BigFloat, Int))
+test_triangular((ComplexF32, ComplexF64, Complex{BigFloat}))
 
-end # module TestTriangularReal
+end # module TestTriangularComplex