… test/testtriag.jl
… test/triangular.jl
… test/triangular2.jl