This brutally splits the triangular tests in two parts, and creates a dedicated test file for unitful tests. That should spread the load more evenly among several workers, when present.

To the best of my knowledge, this doesn't change any test overall. Mostly, this is copy-paste.