@dkarrasch 2024-12-22
The unitful tests take off some heavy load from the bidiag.jl test suite, and the `triangular.jl` tests include only the initial big (nested) loops. All other parts are moved to `triangular2.jl`.