@dkarrasch 2024-12-22
Hm, I missed to check upfront. So the triangular split doesn't yield much, the first block is the very dominant part, so we might as well keep the triangular tests as is. Though any reduction of its runtime is welcome, since this worker runs from start to end and determines the total runtime. I'll check the effect of removing the unitful tests from bidiag.jl tomorrow.