@dkarrasch 2024-12-23
This looks much much better now! Comments welcome.