@dkarrasch 2024-12-24
I'll take a look, but the unitful tests are again dominated by one block of tests (originally from bidiag.jl), like it was the case for triangular.