@dkarrasch 2024-12-24
I skipped the unitless combination in the loop and removed the `Float32` case in the loop. After all, this relies on generic code anyway, so testing through BlasFloats doesn't make much sense. The `Int` case is interesting, because it requires type promotion underneath the `Furlong` coat.