@dkarrasch 2025-01-03
Any objections against this approach?