@KristofferC 2025-01-15
Seems like a good idea to me, the results are nice.