@KristofferC 2025-01-23
I think this broke CI since it needed to be updated after https://github.com/JuliaLang/LinearAlgebra.jl/pull/1152.